\name{as.rglscene}
\alias{as.rglscene}
\title{
Convert an object to an \code{rglscene} object.
}
\description{
This is a placeholder generic function, to allow other
packages to create \code{"rglscene"} objects compatible with
the objects produced by \code{\link{scene3d}}.

No methods are currently defined in \pkg{rgl}.
}
\usage{
as.rglscene(x, ...)
}
\arguments{
  \item{x}{
Object to convert.
}
  \item{\dots}{
Other parameters to pass to methods.
}
}
