\name{persp3d}

\title{
dummy function not real persp3d
}
\description{
I supress rgl
}
\usage{
persp3d(...)
}
\arguments{
  \item{...}{
whatever
}
}

\details{
my sole purpose is to block the wonderful package rgl
}
\value{
none
}

\author{
Soren Havelund Welling, 2016
}

\examples{
persp3d('some stuff')
}
