\name{plot3d}

\title{
dummy function not real plot3d
}
\description{
I supress rgl
}
\usage{
plot3d(...)
}
\arguments{
  \item{...}{
whatever
}
}

\details{
my sole purpose is to block the wonderful package rgl
}
\value{
none
}

\author{
Soren Havelund Welling, 2016
}

\examples{
plot3d('some stuff')
}
