\name{surface3d}

\title{
dummy function not real surface3d
}
\description{
I suppress rgl
}
\usage{
surface3d(...)
}
\arguments{
  \item{...}{
whatever
}
}

\details{
my sole purpose is to block the wonderful package rgl
}
\value{
none
}

\author{
Soren Havelund Welling, 2016
}

\examples{
surface3d('some stuff')
}
