\name{import}
\alias{\%>\%}
\docType{import}
\title{Imported from magrittr}
\description{
This object is imported from \pkg{magrittr}. Follow the link to its documentation.
\describe{
  \item{magrittr}{\code{\link[magrittr:\%>\%]{\%>\%}}}
}

Pipes can be used to string together
\code{\link{rglwidget}} calls and \code{\link{playwidget}}
calls.  See \code{\link{ageControl}} for an example.

}
