\name{in_pkgdown_example}
\alias{in_pkgdown_example}
\title{
Are we running in a \pkg{pkgdown} example?
}
\description{
This is mainly for internal use to decide whether results
should be automatically included in a \pkg{pkgdown} web
page.
}
\usage{
in_pkgdown_example()
}
\value{
\code{TRUE} or \code{FALSE}
}

\examples{
in_pkgdown_example()
}
