\name{rgl-duplicated}
\alias{rgl.pop}
\alias{rgl.cur}
\alias{rgl.ids}
\title{
Duplicated functions in \pkg{rgl}
}
\description{
These functions are provided for compatibility with older versions of
\pkg{rgl}.  See the functions in the comments for the 
full documentation.
}
\usage{
rgl.cur()                             # cur3d      
rgl.ids(type = "shapes", 
        subscene = NA, tags = FALSE)  # ids3d 
rgl.pop(type = "shapes", 
        id = 0, tag = NULL)           # pop3d
}
\keyword{internal}
\keyword{misc}